requirements.txt
# Core Django
Django>=4.2.0,<5.0
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
django-filter>=23.0,<24.0
django-cors-headers>=4.3.0,<5.0

# Database
psycopg2-binary>=2.9.9,<3.0  # PostgreSQL adapter

# Development & Testing
pytest>=7.4.3,<8.0
pytest-django>=4.7.0,<5.0
pytest-cov>=4.1.0,<5.0
black>=23.12.1,<24.0
flake8>=6.1.0,<7.0
isort>=5.13.2,<6.0

# Production
gunicorn>=21.2.0,<22.0
whitenoise>=6.6.0,<7.0

# Utilities
python-decouple>=3.8,<4.0
requests>=2.32.4,<3.0
Pillow>=10.1.0,<11.0  # Image processing
celery>=5.3.4,<6.0  # Task queue
redis>=5.0.1,<6.0  # Cache and message broker

# Monitoring & Logging
sentry-sdk>=1.39.1,<2.0